F #2930: GOCA - split vm file
